+++ pkgs/storage_fs/example/example.dart
@@ -10,7 +10,7 @@ import 'dart:io';
 /// - Work with directories
 void main() async {
   // Initialize storage with configuration
-  Storage.initialize({
+  Storage.initializeFromMap({
     'default': 'local',
     'disks': {
       'local': {